requirements.txt
# Anansi Watchdog Requirements

# Core AI Model APIs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Data Processing
python-dotenv>=1.0.0

# Optional: Enhanced evaluation
# scikit-learn>=1.3.0  # For ML-based evaluation (future)
# transformers>=4.30.0  # For advanced NLP analysis (future)
